+++ app/src/main/java/org/adaway/ui/support/SupportActivity.java
@@ -1,10 +1,14 @@
 package org.adaway.ui.support;
 
+import static android.content.Intent.ACTION_VIEW;
+import static android.net.Uri.parse;
+
 import android.animation.Animator;
 import android.animation.AnimatorListenerAdapter;
 import android.animation.AnimatorSet;
 import android.animation.ObjectAnimator;
 import android.animation.PropertyValuesHolder;
+import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
@@ -16,8 +20,6 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import org.adaway.R;
 
-import static android.content.Intent.ACTION_VIEW;
-
 /**
  * This class is an activity for users to show their supports to the project.
  *
@@ -27,11 +29,11 @@ public class SupportActivity extends AppCompatActivity {
     /**
      * The support link.
      */
-    public static final String SUPPORT_LINK = "https://paypal.me/BruceBUJON";
+    public static final Uri SUPPORT_LINK = parse("https://paypal.me/BruceBUJON");
     /**
      * The sponsorship link.
      */
-    public static final String SPONSORSHIP_LINK = "https://github.com/sponsors/PerfectSlayer";
+    public static final Uri SPONSORSHIP_LINK = parse("https://github.com/sponsors/PerfectSlayer");
 
     public static void animateHeart(ImageView heartImageView) {
         PropertyValuesHolder growScaleX = PropertyValuesHolder.ofFloat(View.SCALE_X, 1F, 1.2F);
@@ -56,32 +58,30 @@ public class SupportActivity extends AppCompatActivity {
         animationSet.start();
     }
 
+    public static void bindLink(Context context, View view, Uri uri) {
+        view.setOnClickListener(v -> {
+            Intent browserIntent = new Intent(ACTION_VIEW, uri);
+            context.startActivity(browserIntent);
+        });
+    }
+
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.support_activity);
 
-        View view = findViewById(R.id.supportRootLayout);
         ImageView heartImageView = findViewById(R.id.headerImageView);
         animateHeart(heartImageView);
-        bindPaypal(view);
-        bindSponsorShip(view);
+        bindPaypal();
+        bindSponsorShip();
     }
 
-    private void bindPaypal(View view) {
-        bindLink(view, R.id.headerImageView, SUPPORT_LINK);
-        bindLink(view, R.id.paypalCardView, SUPPORT_LINK);
+    private void bindPaypal() {
+        bindLink(this, findViewById(R.id.headerImageView), SUPPORT_LINK);
+        bindLink(this, findViewById(R.id.paypalCardView), SUPPORT_LINK);
     }
 
-    private void bindSponsorShip(View view) {
-        bindLink(view, R.id.sponsorshipCardView, SPONSORSHIP_LINK);
-    }
-
-    private void bindLink(View rootView, int viewId, String url) {
-        View view = rootView.findViewById(viewId);
-        view.setOnClickListener(v -> {
-            Intent browserIntent = new Intent(ACTION_VIEW, Uri.parse(url));
-            startActivity(browserIntent);
-        });
+    private void bindSponsorShip() {
+        bindLink(this, findViewById(R.id.sponsorshipCardView), SPONSORSHIP_LINK);
     }
 }